requirements.txt
numpy==1.21.5
pygame==2.1.2
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.5
